requirements.txt
streamlit
python-dotenv
elevenlabs
google-generativeai
